# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MIWAE Corpus Analysis API",
    "description": "A transparent research prototype for candidate-level profiling of textbook functions, grammatical forms, and distribution across corpus segments.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/features": {
      "get": {
        "summary": "Features",
        "operationId": "features_v1_features_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Features V1 Features Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/text": {
      "post": {
        "summary": "Analyze Text",
        "operationId": "analyze_text_v1_analyze_text_post",
        "parameters": [
          {
            "name": "include_hits",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Hits"
            }
          },
          {
            "name": "max_hits",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 0,
              "default": 5,
              "title": "Max Hits"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Analyze Text V1 Analyze Text Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/corpus": {
      "post": {
        "summary": "Analyze Corpus",
        "operationId": "analyze_corpus_v1_analyze_corpus_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorpusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Analyze Corpus V1 Analyze Corpus Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyze/file": {
      "post": {
        "summary": "Analyze File",
        "operationId": "analyze_file_v1_analyze_file_post",
        "parameters": [
          {
            "name": "include_hits",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Hits"
            }
          },
          {
            "name": "max_hits",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 0,
              "default": 5,
              "title": "Max Hits"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_file_v1_analyze_file_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Analyze File V1 Analyze File Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/adapters/textinspector": {
      "post": {
        "summary": "Import Textinspector",
        "operationId": "import_textinspector_v1_adapters_textinspector_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextInspectorImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Import Textinspector V1 Adapters Textinspector Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/adapters/reference-map": {
      "post": {
        "summary": "Import Reference Map",
        "operationId": "import_reference_map_v1_adapters_reference_map_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceMappingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Import Reference Map V1 Adapters Reference Map Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_analyze_file_v1_analyze_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_file_v1_analyze_file_post"
      },
      "CorpusRequest": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Documents"
          },
          "include_hits": {
            "type": "boolean",
            "title": "Include Hits",
            "default": true
          },
          "max_hits_per_feature": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Max Hits Per Feature",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "CorpusRequest"
      },
      "DocumentInput": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentInput"
            },
            "type": "array",
            "title": "Segments"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "external_profiles": {
            "additionalProperties": true,
            "type": "object",
            "title": "External Profiles"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DocumentInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReferenceMappingRequest": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ReferenceMappingRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "ReferenceMappingRequest"
      },
      "ReferenceMappingRow": {
        "properties": {
          "feature_id": {
            "type": "string",
            "title": "Feature Id"
          },
          "cefr_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cefr Level"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "feature_id"
        ],
        "title": "ReferenceMappingRow"
      },
      "SegmentInput": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "eligible_opportunities": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Eligible Opportunities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text"
        ],
        "title": "SegmentInput"
      },
      "TextInspectorImportRequest": {
        "properties": {
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "TextInspectorImportRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}